# internal/docs/swagger.yaml
basePath: /
definitions:
  github_com_kimnt93_gorouter_internal_api.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
      type:
        type: string
    type: object
  github_com_kimnt93_gorouter_internal_api.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorDetail'
    type: object
...
  github_com_kimnt93_gorouter_pkg_entities.OrganizationModel:
    properties:
      created_at:
        type: string
      enabled:
        type: boolean
      kind:
        type: string
      name:
        type: string
      organization_id:
        type: string
      source_owner_id:
        type: string
      targets:
        items:
          type: string
        type: array
      updated_at:
        type: string
      weekly_limit_usd:
        type: number
    type: object
  github_com_kimnt93_gorouter_pkg_entities.OrganizationModelGrant:
    properties:
      enabled:
        type: boolean
      model:
        type: string
      organization_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
      weekly_limit_usd:
        type: number
    type: object
...
  internal_api_handlers.OrganizationGrantRequest:
    properties:
      enabled:
        type: boolean
      model:
        type: string
      user_id:
        type: string
      weekly_limit_usd:
        type: number
    type: object
...
  internal_api_handlers.OrganizationModelRequest:
    properties:
      enabled:
        type: boolean
      kind:
        type: string
      name:
        type: string
      targets:
        items:
          type: string
        type: array
      weekly_limit_usd:
        type: number
    type: object
...
info:
  contact: {}
  description: Principal-aware model gateway and administration API.
  title: gorouter API
  version: "1.0"
paths:
...
  /admin/organizations/{id}/model-grants:
    get:
      consumes:
      - application/json
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: Assignment; POST
        in: body
        name: request
        schema:
          $ref: '#/definitions/internal_api_handlers.OrganizationGrantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_kimnt93_gorouter_pkg_entities.OrganizationModelGrant'
            type: array
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_pkg_entities.OrganizationModelGrant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List or assign organization models to users
      tags:
      - organization-models
    post:
      consumes:
      - application/json
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: Assignment; POST
        in: body
        name: request
        schema:
          $ref: '#/definitions/internal_api_handlers.OrganizationGrantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_kimnt93_gorouter_pkg_entities.OrganizationModelGrant'
            type: array
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_pkg_entities.OrganizationModelGrant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List or assign organization models to users
      tags:
      - organization-models
  /admin/organizations/{id}/models:
    get:
      consumes:
      - application/json
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: Alias or group; POST
        in: body
        name: request
        schema:
          $ref: '#/definitions/internal_api_handlers.OrganizationModelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_kimnt93_gorouter_pkg_entities.OrganizationModel'
            type: array
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_pkg_entities.OrganizationModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List or publish organization models
      tags:
      - organization-models
    post:
      consumes:
      - application/json
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: Alias or group; POST
        in: body
        name: request
        schema:
          $ref: '#/definitions/internal_api_handlers.OrganizationModelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_kimnt93_gorouter_pkg_entities.OrganizationModel'
            type: array
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_pkg_entities.OrganizationModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List or publish organization models
      tags:
      - organization-models
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Use "Bearer {master-or-api-key}".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"